// src/generated/avoma.openapi.ts
export interface paths {
// ...
    "/v1/custom_categories/{uuid}/": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Custom Category
         * @description Returns a single custom category for a given `uuid`.
         */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path: {
                    uuid: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /** @description AV field mapping of the custom category. */
                            av_field_mapping?: Record<string, never>;
                            /** @description Background color of the custom category. */
                            background_color?: string;
                            /** @description Is default custom category. */
                            is_default?: boolean;
                            /** @description Key of the custom category. */
                            key?: string;
                            /** @description Keywords of the custom category. */
                            keywords?: string[];
                            /** @description Name of the custom category. */
                            name?: string;
                            /** @description Order of the custom category. */
                            order?: number;
                            /** @description Prompts of the custom category. */
                            prompts?: string[];
                            /** @description Settings of the custom category. */
                            settings?: {
                                /** @description Augmented notes enabled. */
                                aug_notes_enabled?: boolean;
                                /** @description Keyword notes enabled. */
                                keyword_notes_enabled?: boolean;
                                /** @description Keyword tracking enabled. */
                                keyword_tracking_enabled?: boolean;
                                /** @description Meeting type rule. */
                                meeting_type_rule?: number;
                                /** @description Meeting types blacklisted. */
                                meeting_types_blacklisted?: string[];
                                /** @description Meeting types whitelisted. */
                                meeting_types_whitelisted?: string[];
                                /** @description Prompt extract length. */
                                prompt_extract_length?: string;
                                /** @description Prompt extract strategy. */
                                prompt_extract_strategy?: string;
                                /** @description Prompt notes enabled. */
                                prompt_notes_enabled?: boolean;
                            };
                            /**
                             * Format: uuid
                             * @description UUID of the custom category.
                             */
                            uuid?: string;
                        };
                    };
                };
                /** @description Not Found. */
                404: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /**
                             * @description Error details
                             * @example Not Found
                             */
                            detail: string;
                        };
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/v1/meeting_segments/": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get Meeting segments
         * @description Returns segments of the meeting for a given `uuid`.
         */
        get: {
            parameters: {
                query: {
                    uuid: string;
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            agenda?: number[][];
                            demo?: number[][];
                            intro?: number[][];
                            meeting?: number[][];
                            next_steps?: number[][];
                            objection?: number[][];
                            overview?: number[][];
                            pain_point?: number[][];
                            pricing?: number[][];
                        };
                    };
                };
                /** @description Not Found. */
                404: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /**
                             * @description Error details
                             * @example Not Found
                             */
                            detail: string;
                        };
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v1/notes/": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * List Notes
         * @description Returns all notes from the organization if no params are passed. If `meeting` is passed, returns all notes for that meeting. If `custom_category` is passed, returns all notes for that custom category. If `output_format` is passed, returns all notes for that output format. Choices are `json`,`html`,`markdown` ordering can be done by `created` or `modified` by passing `o=<field>` in params
         */
        get: {
            parameters: {
                query: {
                    /** @description Number of records returned per response. */
                    page_size?: number;
                    /** @description To retrieve notes for meetings started at or started after given date-time(UTC). */
                    from_date: string;
                    /** @description To retrieve notes for meetings started at or started before given date-time(UTC). */
                    to_date: string;
                    o?: "start_at" | "-start_at" | "modified" | "-modified";
                    output_format?: "json" | "html" | "markdown";
                    meeting_uuid?: string;
                    custom_category?: string;
                };
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description OK */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /** @description Total number of records in result. */
                            count?: number;
                            /**
                             * @description URL to the next page if available.
                             * @example https://example.com/api/v1/meeting/?page=2
                             */
                            next?: string | null;
                            /**
                             * @description URL to the previous page if available.
                             * @example https://example.com/api/v1/meeting/?page=1
                             */
                            previous?: string | null;
                        } & {
                            /** @description Actual response. */
                            results: {
                                /**
                                 * Format: date-time
                                 * @description Last modification date-time(UTC).
                                 */
                                created?: string;
                                /** @description Notes data. */
                                data?: Record<string, never> | Record<string, never>[] | string;
                                /**
                                 * Format: date-time
                                 * @description Last modification date-time(UTC).
                                 */
                                modified?: string;
                            }[];
                        };
                    };
                };
                /** @description No content data found */
                204: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /**
                             * @description Error details
                             * @example Not valid data found
                             */
                            detail: string;
                        };
                    };
                };
                /** @description Invalid query params */
                400: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            /**
                             * @description Error details
                             * @example Not Found
                             */
                            detail: string;
                        };
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/v1/scorecard_evaluations/": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * List Scorecard Evaluations
         * @description Retrieve a list of scorecard evaluations for meetings.
         */
        get: operations["scorecard_evaluations_list"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}